# Store `G.adj` as a local variable to speed up `complement_edges(G)`