--- a/networkx/algorithms/connectivity/edge_augmentation.py
+++ b/networkx/algorithms/connectivity/edge_augmentation.py
@@ -1124,16 +1124,16 @@ def complement_edges(G):
     >>> sorted(complement_edges(G))
     []
     """
-    adj = G._adj
+    G_adj = G._adj  # Store as a variable to eliminate attribute lookup
     if G.is_directed():
         for u, v in it.combinations(G.nodes(), 2):
-            if v not in adj[u]:
+            if v not in G_adj[u]:
                 yield (u, v)
-            if u not in adj[v]:
+            if u not in G_adj[v]:
                 yield (v, u)
     else:
         for u, v in it.combinations(G.nodes(), 2):
-            if v not in adj[u]:
+            if v not in G_adj[u]:
                 yield (u, v)